Renamed the function from the basic hello world function. The local build and invoke still working
