+++ machineLearning/sentimentAnalysis/newsSentimentAnalysis/NewsSentimentLambdaFunction/.aws-sam/build.toml
@@ -1,13 +1,13 @@
 # This file is auto generated by SAM CLI build command
 
 [function_build_definitions]
-[function_build_definitions.23dc33ee-cdd3-481d-a6e2-e9021f218f82]
+[function_build_definitions.0bd1089a-b14d-4e76-8e40-e50f65fb1b13]
 packagetype = "Image"
-functions = ["HelloWorldFunction"]
+functions = ["NewsSentimentFunction"]
 
-[function_build_definitions.23dc33ee-cdd3-481d-a6e2-e9021f218f82.metadata]
+[function_build_definitions.0bd1089a-b14d-4e76-8e40-e50f65fb1b13.metadata]
 Dockerfile = "Dockerfile"
-DockerContext = "C:\\Users\\warre\\OneDrive\\Documents\\College\\Msc\\FinalYearProject\\AWS\\Amplify\\stockApp\\machineLearning\\sentimentAnalysis\\newsSentimentAnalysis\\NewsSentimentLambdaFunction\\hello_world"
+DockerContext = "C:\\Users\\warre\\OneDrive\\Documents\\College\\Msc\\FinalYearProject\\AWS\\Amplify\\stockApp\\machineLearning\\sentimentAnalysis\\newsSentimentAnalysis\\NewsSentimentLambdaFunction\\newsSentiment"
 DockerTag = "python3.9-v1"
 
 [layer_build_definitions]
