+++ machineLearning/sentimentAnalysis/newsSentimentAnalysis/NewsSentimentLambdaFunction/newsSentiment/app.py
