+++ machineLearning/sentimentAnalysis/newsSentimentAnalysis/NewsSentimentLambdaFunction/newsSentiment/requirements.txt
