+++ machineLearning/sentimentAnalysis/newsSentimentAnalysis/NewsSentimentLambdaFunction/tests/unit/test_handler.py
@@ -2,7 +2,7 @@ import json
 
 import pytest
 
-from hello_world import app
+from newsSentiment import app
 
 
 @pytest.fixture()
